+++ gba_test_macros/tests/trybuild/ignore_invalid_arguments.rs
@@ -1,0 +1,10 @@
+#![feature(custom_test_frameworks)]
+
+use gba_test_macros::test;
+
+#[test]
+#[ignore("foo")]
+fn foo() {}
+
+
+fn main() {}