+++ gba_test_macros/tests/trybuild/should_panic_arguments_not_name_value.rs
@@ -1,0 +1,10 @@
+#![feature(custom_test_frameworks)]
+
+use gba_test_macros::test;
+
+#[test]
+#[should_panic("foo")]
+fn foo() {}
+
+
+fn main() {}