+++ gba_test_macros/tests/trybuild/should_panic_incorrect_argument_name.rs
@@ -1,0 +1,10 @@
+#![feature(custom_test_frameworks)]
+
+use gba_test_macros::test;
+
+#[test]
+#[should_panic(expectd = "foo")]
+fn foo() {}
+
+
+fn main() {}